python-common: Initialize package structure
This package is supposed to contain common Python code usable by all modules and tools.

It is also supposed to contain code to deploy ceph clutsers.
