+++ src/python-common/.gitignore
@@ -1,0 +1,1 @@
+ceph.egg-info
