+++ src/python-common/ceph/deployment/ssh_orchestrator.py
@@ -1,0 +1,10 @@
+
+def bootstrap_cluster():
+    create_mon()
+    create_mgr()
+
+def create_mon():
+    pass
+
+def create_mgr():
+    pass
